fix: improve form error styling, consolidate state management, and clean up actions organization
