+++ front_end/src/app/(campaigns-registration)/rand/constants.ts
@@ -1,2 +1,1 @@
-export const CAMPAIGN_KEY = "rand_forecasting_2025";
 export const CAMPAIGN_URL_BASE_PATH = "/rand";
